Move timer slice to slices folder and clean up some linting issues
