+++ orchestra/frontend/src/state/rootReducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from '@reduxjs/toolkit'
 
 import dashboardTasks from './slices/dashboardTasks'
-import timer from './timer'
+import timer from './slices/timer'
 
 const rootReducer = combineReducers({
   [dashboardTasks.name]: dashboardTasks.reducer,
